+++ sdk/web-pubsub/web-pubsub-client/test/asyncSeqQueue.spec.ts
@@ -62,27 +62,27 @@ describe("AsyncSeqQueue", () => {
     await expect(queue.dequeue()).resolves.toEqual({ sequenceId: 3, value: "C" });
   });
 
-  it("clamps reset to unacked range", async () => {
+  it("throws when reset sequence id is older than oldest unacked", async () => {
     const queue = new AsyncSeqQueue<string>();
     await queue.enqueue("A");
     await queue.enqueue("B");
     await queue.enqueue("C");
     queue.ack(3);
 
-    queue.reset(1);
+    expect(() => queue.reset(1)).toThrow("Invalid sequence id 1 for reset");
     expect(queue.nextDequeueSequenceId).toBe(3);
     await expect(queue.dequeue()).resolves.toEqual({ sequenceId: 3, value: "C" });
   });
 
-  it("clamps ack to latestEnqueued + 1", async () => {
+  it("throws when ack sequence id is larger than next enqueue sequence id", async () => {
     const queue = new AsyncSeqQueue<string>();
     await queue.enqueue("A");
     await queue.enqueue("B");
 
-    expect(queue.ack(100)).toBe(2);
-    expect(queue.size).toBe(0);
-    expect(queue.oldestUnackedSequenceId).toBe(3);
-    expect(queue.nextDequeueSequenceId).toBe(3);
+    expect(() => queue.ack(100)).toThrow("Invalid sequence id 100 for ack");
+    expect(queue.size).toBe(2);
+    expect(queue.oldestUnackedSequenceId).toBe(1);
+    expect(queue.nextDequeueSequenceId).toBe(1);
   });
 
   it("supports custom initial sequence ids", async () => {
@@ -101,30 +101,27 @@ describe("AsyncSeqQueue", () => {
     await expect(queue.dequeue()).resolves.toEqual({ sequenceId: 11, value: "B" });
   });
 
-  it("ignores ack that is older than the oldest unacked sequence id", async () => {
+  it("throws when ack sequence id is older than the oldest unacked sequence id", async () => {
     const queue = new AsyncSeqQueue<string>();
     await queue.enqueue("A");
     await queue.enqueue("B");
 
     expect(queue.ack(2)).toBe(1);
-    expect(queue.ack(1)).toBe(0);
+    expect(() => queue.ack(1)).toThrow("Invalid sequence id 1 for ack");
     expect(queue.size).toBe(1);
     expect(queue.oldestUnackedSequenceId).toBe(2);
     expect(queue.nextDequeueSequenceId).toBe(2);
     await expect(queue.dequeue()).resolves.toEqual({ sequenceId: 2, value: "B" });
   });
 
-  it("clamps reset above the latest enqueued sequence id to the next enqueue sequence id", async () => {
+  it("throws when reset sequence id is larger than next enqueue sequence id", async () => {
     const queue = new AsyncSeqQueue<string>();
     await queue.enqueue("A");
     await queue.enqueue("B");
 
-    queue.reset(100);
-    expect(queue.nextDequeueSequenceId).toBe(3);
-
-    const pending = queue.dequeue();
-    await expect(queue.enqueue("C")).resolves.toBe(3);
-    await expect(pending).resolves.toEqual({ sequenceId: 3, value: "C" });
+    expect(() => queue.reset(100)).toThrow("Invalid sequence id 100 for reset");
+    expect(queue.nextDequeueSequenceId).toBe(1);
+    await expect(queue.dequeue()).resolves.toEqual({ sequenceId: 1, value: "A" });
   });
 
   it("serves multiple pending dequeue waiters in order", async () => {
